Fixed float formatting

+++ Assets/Geo_Quest/Scripts/randomThing.cs
@@ -4,7 +4,7 @@ using Unity.Burst.Intrinsics;
 using UnityEngine;
 public class randomThing : MonoBehaviour
 {
-    float thing = (float)1.0;
+    float thing = 1.0f;
     string variable1 = "Hello";
     private int invert = -1;
     // Start is called before the first frame update
@@ -18,7 +18,7 @@ public class randomThing : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        thing *= (float)1.01;
+        thing *= 1.01f;
         thing *= invert;
         Debug.Log(transform.position);
         transform.position = new Vector3(thing, thing, thing);
